checkbox updated reducer action
