+++ src/components/DiscountedCheckBox/index.jsx
@@ -1,4 +1,31 @@
 
+// import React from 'react';
+// import defaultStyles from './DiscountedCheckBox.module.css';
+// import checkedIcon from '../../media/icons/check box=active.svg'; 
+// import uncheckedIcon from '../../media/icons/check box=normal.svg';
+
+// const DiscountedCheckBox = ({ styles = defaultStyles, setShowDiscounted, showDiscounted }) => {
+//   const handleCheckboxClick = () => {
+//     setShowDiscounted(!showDiscounted);
+//   };
+
+//   return (
+//     <div className={styles.discount_wrapper}>
+//       <label htmlFor="discounted"><h3>Discounted items</h3></label>
+//       <div className={styles.checkbox} onClick={handleCheckboxClick}>
+//         <img
+//           src={showDiscounted ? checkedIcon : uncheckedIcon} 
+//           className={styles.checkbox_image}
+//              alt="Discount checkbox"
+//         />
+//       </div>
+//     </div>
+//   );
+// };
+
+// export default DiscountedCheckBox;
+
+
 import React from 'react';
 import defaultStyles from './DiscountedCheckBox.module.css';
 import checkedIcon from '../../media/icons/check box=active.svg'; 
@@ -16,7 +43,7 @@ const DiscountedCheckBox = ({ styles = defaultStyles, setShowDiscounted, showDis
         <img
           src={showDiscounted ? checkedIcon : uncheckedIcon} 
           className={styles.checkbox_image}
-             alt="Discount checkbox"
+          alt="Discount checkbox"
         />
       </div>
     </div>
