+++ src/components/ProductDetail/index.jsx
@@ -1,3 +1,4 @@
+
 import React, { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
@@ -113,7 +114,7 @@ const ProductDetail = () => {
                             handleDecrement={handleDecrement}
                         />
                         <ButtonAddToCard
-                        className={styles.button_add}
+                            className={styles.button_add}
                             product={{ ...product, count: productCount }}
                         />
                     </div>
