+++ src/redux/actionTypes.js
@@ -1,0 +1,20 @@
+export const actionTypes = {
+    FETCH_PRODUCTS_REQUEST: 'FETCH_PRODUCTS_REQUEST',
+    FETCH_PRODUCTS_SUCCESS: 'FETCH_PRODUCTS_SUCCESS',
+    FETCH_PRODUCTS_FAILURE: 'FETCH_PRODUCTS_FAILURE',
+    FETCH_PRODUCTS_BY_CATEGORY_REQUEST: 'FETCH_PRODUCTS_BY_CATEGORY_REQUEST',
+    FETCH_PRODUCTS_BY_CATEGORY_SUCCESS: 'FETCH_PRODUCTS_BY_CATEGORY_SUCCESS',
+    FETCH_PRODUCTS_BY_CATEGORY_FAILURE: 'FETCH_PRODUCTS_BY_CATEGORY_FAILURE',
+    FETCH_PRODUCT_REQUEST: 'FETCH_PRODUCT_REQUEST',
+    FETCH_PRODUCT_SUCCESS: 'FETCH_PRODUCT_SUCCESS',
+    FETCH_PRODUCT_FAILURE: 'FETCH_PRODUCT_FAILURE',
+    INCREMENT_PRODUCT_COUNT: 'INCREMENT_PRODUCT_COUNT',
+    DECREMENT_PRODUCT_COUNT: 'DECREMENT_PRODUCT_COUNT',
+    ADD_TO_FAVORITES: 'ADD_TO_FAVORITES',
+    REMOVE_FROM_FAVORITES: 'REMOVE_FROM_FAVORITES',
+    SET_MIN_PRICE_FILTER: 'SET_MIN_PRICE_FILTER',
+    SET_MAX_PRICE_FILTER: 'SET_MAX_PRICE_FILTER',
+    SET_SORT_BY: 'SET_SORT_BY',
+    SET_SHOW_DISCOUNTED: 'SET_SHOW_DISCOUNTED',
+    SET_FILTERED_AND_SORTED_PRODUCTS: 'SET_FILTERED_AND_SORTED_PRODUCTS',
+  };
